fully optimized the home page using ssr
